requirements.txt
asgiref==3.2.7
Django==3.0.4
django-filter==2.2.0
django-tawkto==0.2
Pillow==7.1.1
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
waitress==2.0.0
whitenoise==5.2.0
